move normalize_path to top level
