+++ core/lib.rs
@@ -15,11 +15,11 @@ mod es_isolate;
 mod flags;
 mod module_specifier;
 mod modules;
+mod normalize_path;
 mod ops;
 pub mod plugin_api;
 mod resources;
 mod shared_queue;
-pub mod utils;
 mod zero_copy_buf;
 
 pub use rusty_v8 as v8;
@@ -44,6 +44,7 @@ pub use crate::modules::ModuleLoader;
 pub use crate::modules::ModuleSource;
 pub use crate::modules::ModuleSourceFuture;
 pub use crate::modules::RecursiveModuleLoad;
+pub use crate::normalize_path::normalize_path;
 pub use crate::ops::Buf;
 pub use crate::ops::Op;
 pub use crate::ops::OpAsyncFuture;
