+++ core/normalize_path.rs
